fix(otel): resolve lint and CI test coverage for Azure OTel hooks
Fix eslint violations, consolidate unit tests under test/helpers so
verify-exercised-tests picks them up via test:instrumentations:misc.
